# server/docs/swagger-bundled.yaml
openapi: 3.0.0
info:
  title: Users API
  version: 1.0.0
  description: API for managing users with HAL-style hypermedia responses
servers:
  - url: 'http://localhost:3001'
paths:
...
  /api/items:
    get:
      summary: Get all items
      description: Retrieve a list of all items from the database.
      tags:
        - All Items
      responses:
        '200':
          description: A list of items
          content:
            application/json:
              schema:
                type: object
                properties:
                  _links:
                    type: object
                    properties:
                      self:
                        type: object
                        properties:
                          href:
                            type: string
                            example: /api/items/
                      profile:
                        type: object
                        properties:
                          href:
                            type: string
                            example: /profile/items/
                      create:
                        type: object
                        properties:
                          href:
                            type: string
                            example: /api/items
                          method:
                            type: string
                            example: POST
                  _embedded:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Item'
        '204':
          description: Items not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Items not found
        '500':
          $ref: '#/components/schemas/InternalServerError'
    post:
      summary: Add a new item
      description: Create a new item in the database.
      tags:
        - All Items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddItemRequest'
      responses:
        '201':
          description: Item created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _links:
                    type: object
                    properties:
                      self:
                        type: object
                        properties:
                          href:
                            type: string
                            example: /api/items/1
                      all:
                        type: object
                        properties:
                          href:
                            type: string
                            example: /api/items
                      profile:
                        type: object
                        properties:
                          href:
                            type: string
                            example: /profile/items
                      edit:
                        type: object
                        properties:
                          href:
                            type: string
                            example: /api/items/1
                          method:
                            type: string
                            example: PUT
                      delete:
                        type: object
                        properties:
                          href:
                            type: string
                            example: /api/items/1
                          method:
                            type: string
                            example: DELETE
                  dataValues:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: New China Sample Item
                      description:
                        type: string
                        example: Ergonomic Chicken made with Silk for all-day oblong support
                      seller_id:
                        type: integer
                        example: 1
                  _previousDataValues:
                    type: object
                    description: Same structure as dataValues
                    properties:
                      name:
                        type: string
                        example: Sample Item
                      description:
                        type: string
                        example: Ergonomic Chicken made with Silk for all-day oblong support.
                      seller_id:
                        type: integer
                        example: 1
                      id:
                        type: integer
                        example: 1
                  uniqno:
                    type: integer
                    example: 1
                  _changed:
                    type: object
                    example: {}
                  _options:
                    type: object
                    properties:
                      isNewRecord:
                        type: boolean
                        example: true
                      _schema:
                        type: string
                        nullable: true
                        example: null
                      _schemaDelimiter:
                        type: string
                        example: ''
                  isNewRecord:
                    type: boolean
                    example: false
        '400':
          $ref: '#/components/schemas/BadRequest'
        '500':
          $ref: '#/components/schemas/InternalServerError'
...
  /profile/items:
    get:
      tags:
        - Profile
      summary: Retrieve Items profile information
      description: Fetches the profile details and actions for items.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemsProfile'
        '404':
          description: Items Profile not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Items Profile not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Item:
      type: object
      properties:
        _links:
          type: object
          properties:
            self:
              type: object
              properties:
                href:
                  type: string
                  example: /api/items/1
            edit:
              type: object
              properties:
                href:
                  type: string
                  example: /api/items/1
                method:
                  type: string
                  example: PUT
            delete:
              type: object
              properties:
                href:
                  type: string
                  example: /api/items/1
                method:
                  type: string
                  example: DELETE
        dataValues:
          type: object
          properties:
            id:
              type: integer
              example: 1
            name:
              type: string
              example: Sample Item
            description:
              type: string
              example: Ergonomic Chicken made with Silk for all-day oblong support.
            sellerId:
              type: integer
              example: 1
            categoryId:
              type: integer
              example: 2
            seller_id:
              type: integer
              example: 1
        _previousDataValues:
          type: object
          description: Same structure as dataValues
          properties:
            id:
              type: integer
              example: 1
            name:
              type: string
              example: Sample Item
            description:
              type: string
              example: Ergonomic Chicken made with Silk for all-day oblong support.
            sellerId:
              type: integer
              example: 1
            categoryId:
              type: integer
              example: 2
            seller_id:
              type: integer
              example: 1
        uniqno:
          type: integer
          example: 1
        _changed:
          type: object
          example: {}
        _options:
          type: object
          properties:
            isNewRecord:
              type: boolean
              example: false
            _schema:
              type: string
              nullable: true
              example: null
            _schemaDelimiter:
              type: string
              example: ''
            raw:
              type: boolean
              example: true
            attributes:
              type: array
              items:
                type: string
                example:
                  - id
                  - name
                  - description
                  - sellerId
                  - categoryId
                  - seller_id
        isNewRecord:
          type: boolean
          example: false
    AddItemRequest:
      type: object
      required:
        - name
        - description
        - sellerId
        - categoryId
      properties:
        name:
          type: string
          example: New Chinese Tariff Item
        description:
          type: string
          example: This is a US 125% on Items
        seller_id:
          type: integer
          example: 1
        category_id:
          type: integer
          example: 2
...
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid request body
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: An unexpected error occurred
...
    ItemsProfile:
      type: object
      properties:
        _links:
          type: object
          properties:
            self:
              type: object
              properties:
                href:
                  type: string
                  example: /profile/items
            item:
              type: object
              properties:
                href:
                  type: string
                  example: '/api/items/:id'
                templated:
                  type: boolean
                  example: true
        properties:
          type: object
          properties:
            id:
              type: object
              properties:
                type:
                  type: string
                  example: integer
                description:
                  type: string
                  example: The unique identifier for the item.
            name:
              type: object
              properties:
                type:
                  type: string
                  example: string
                description:
                  type: string
                  example: The name of the item.
                maxLength:
                  type: integer
                  example: 100
            description:
              type: object
              properties:
                type:
                  type: string
                  example: string
                description:
                  type: string
                  example: A detailed description of the item.
            user_id:
              type: object
              properties:
                type:
                  type: string
                  example: integer
                description:
                  type: string
                  example: The unique identifier for the user who owns the item.
                format:
                  type: string
                  example: int64
            category_id:
              type: object
              properties:
                type:
                  type: string
                  example: integer
                description:
                  type: string
                  example: The unique identifier for the category the item belongs to.
                format:
                  type: string
                  example: int64
        actions:
          type: array
          items:
            oneOf:
              - type: object
                properties:
                  name:
                    type: string
                    example: create
                  method:
                    type: string
                    example: POST
                  href:
                    type: string
                    example: /api/items
                  title:
                    type: string
                    example: Create a new object to items
              - type: object
                properties:
                  name:
                    type: string
                    example: get all
                  method:
                    type: string
                    example: GET
                  href:
                    type: string
                    example: /api/items
                  title:
                    type: string
                    example: Get all objects from items
              - type: object
                properties:
                  name:
                    type: string
                    example: get single
                  method:
                    type: string
                    example: GET
                  href:
                    type: string
                    example: '/api/items/:id'
                  title:
                    type: string
                    example: 'Get a single object from items by :id'
              - type: object
                properties:
                  name:
                    type: string
                    example: delete
                  method:
                    type: string
                    example: DELETE
                  href:
                    type: string
                    example: '/api/items/:id'
                  templated:
                    type: boolean
                    example: true
                  title:
                    type: string
                    example: Delete object from items
              - type: object
                properties:
                  name:
                    type: string
                    example: update
                  method:
                    type: string
                    example: PUT
                  href:
                    type: string
                    example: '/api/items/:id'
                  templated:
                    type: boolean
                    example: true
                  title:
                    type: string
                    example: Update an existing object in items
...